requirements.txt
APScheduler
environs
psycopg2-binary
python-telegram-bot
sqlalchemy
web3
